pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supernote-to-obsidian"
version = "0.1.0"
description = "Sync handwritten notes from a Supernote Manta tablet to an Obsidian vault via OCR"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Apache-2.0" }
license-files = ["LICENSE"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "watchdog>=4.0",
    "Pillow>=10",
    "pytesseract>=0.3.10",
    "google-cloud-vision>=3.7",
    "python-frontmatter>=1.1",
    "PyYAML>=6.0",
    "click>=8.1",
    "APScheduler>=3.10",
    "rich>=13.7",
    "requests>=2.31",
    "pdf2image>=1.17",
]

[project.optional-dependencies]
supernote = [
    "supernote-tool>=0.1",
]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "ruff",
    "mypy",
]

[project.scripts]
supernote-sync = "supernote_sync.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["supernote_sync*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=supernote_sync --cov-report=term-missing"

[tool.coverage.report]
fail_under = 80
